Make all methods async
This makes this module non-blocking entirely. We cannot call the
`authenticate()` from init anymore since init cannot be async, but I
think that is a cost we can eat, any module that uses this should be
calling authenticate before doing anything anyways.

This gives us the benefit that in usages like a discord bot, it won't
block the thread on these web requests, we can let it take its time and
run other things while we wait for responses.