Add basic logging
The parent application may want to pass in a more complex logger, but we
should have some sort of logging regardless. We should also add logging
all over the place so we can follow when things pass and fail so we get
a better picture of what is happening when we run this on prod.