Add retry on invalid cookie or not logged in
This allows any of the get methods to try again after a login call to
make sure that we will not keep trying with an invalid cookie. iRacing
cookies fail every ~12 hours so we would not expect this to happen
frequently.

This also raises a RuntimeError if the call is tried again and fails the
2nd time as well, because if that happens something is very wrong and we
either need human intervention, or we need to wait for the servers to
come back up :).