Add methods for irating, ttrating and license class
Generally people will not want to call `ChartData` because it is unclear
exactly what data it gives you, so these methods get the relevant data
out of the ChartData.

Each of these methods returns a single ChartData object, each of which
contain a list of either iratings, ttratings, or license classes.