Add parameters for current_seasons call
We want to allow the user to set the params being passed in here even if
it is a complete mess. This also defines a method to build us the list
we need to make it cleaner in case we find another endpoint we need to
do the same thing for later on.