Improve helper methods on constants
In these constants we want to be able to get a readable string from a
number that represents a value in each one. By forcing them to be
instance methods, it means I can check based on the name of the variable
instead of just the number, so if iRacing decides that rookie is now 0
instead of 1, we just change the mapping for rookie and the method is
already up to date, we can't forget to update it and let it break.