Add object return for current_seasons endpoint
This should make the current_seasons endpoint more usable. It takes the
return and maps it into a list of season objects. The Season object only
forces the fields `seasonId`, `catid` and `seriesshortname` to be
present, the rest can be None because the endpoint won't return those
unless we explicitly ask for them.