Fix some syntax errors
These are a few errors that came up from a previous branch. The `n`
looks like a typo, when comparing to a `response.url` that returns a url
object, which you cannot iterate over to check if something is in it, so
this needs to be the string of the URL, not the URL object. The constant
`SimSesNum.RACE` was just another typo as this variable is in
`SimSesNum.race`.