Add testing to pyracing
While this is a pretty hacky test that doesn't do as much as I would
like, it will at least enforce that the methods we use return the type
of object we expect. If we know that these objects are getting created,
then we know that we are getting the correct info back because building
the objects would explode otherwise. This should help us be safer when
making changes.