Simplify logic for 0 indexed race_week
race_week is 0 indexed, meaning week 1 needs to be passed as 0, so we
want to user interface to be good and allow users to pass 1 for week 1.
We also need to make sure if they pass -1 that it stays because that is
a valid input but -2 is not.

Also taking this opportunity to update some field names to proper
snake_case.