Auto set current year and require quarter in event_results
The year is easy enough to make current, but the quarter is a bit more
work, so we will let the user supply it for now while we find an
effective way to get the current quarter without making extra network
calls.

This also moves the `datetime_from_iracing_timestamp` to outside of
chart data, this converts unix timestamps from ms to datetimes and we
will be using this in more places than just the chart data objects.