Use data class for chart data classes
This makes these objects read a lot cleaner and makes them easier to
understand. It uses some inheritance for the timestamp and datetime
stuff that is in all of the different chart data content classes.

This also changes how we store this data, since they send a timestamp
and we parse it into a datetime we should just save the timestamp and
only get the datetime when we call the method. This lets us not explode
if the timestamp is not parseable, it will only explode on the call to
`datetime()`.