Improve variable names and error reporting
We previously had the help message inside the error we are raising, but
we really want to be logging that as that is where users will be looking
for help. Inside the error we can just dump the entire response. This
might be absurd and too much content, but it at least allows them to
make decisions based on this if they want.