New client method to rename numerical keys
Takes advantage of the mapping iracing returns alongside the data with numerical keys. This should be a future-proof against breaking or importing the wrong data, when iracing changes the numerical keys the next time.

All client methods that deal with numerically keyed data have been adapted to use this.

The response_object classes affected by this have also been updated.